C-plus-plus Merged
