+++ C-plus-plus/Number-Programs/Armstrong.cpp
