+++ C-plus-plus/Number-Programs/Niven-number-or-Harshad-number.cpp
