+++ C-plus-plus/Patterns/Binary-Pattern.cpp
