+++ C-plus-plus/Patterns/Diamond-Pattern.cpp
