+++ C-plus-plus/Patterns/Diamond-Star-Outline.cpp
