+++ C-plus-plus/Patterns/Diamond-of-Numbers.cpp
