+++ C-plus-plus/Patterns/Number-Diamond-Pattern.cpp
