+++ C-plus-plus/Patterns/Pyramid-Pattern-2.cpp
